refactor: parser.cpp と command_generator.cpp を責務別に分割 (Phase 2)
- parser.cpp (1239行→975行): Alert検出を parser_alerts.cpp、HTMLエンティティを html_entities.cpp/h へ分離
- command_generator.cpp (967行→637行): テーブル描画を command_generator_table.cpp、選択・検索ハイライトを command_generator_highlights.cpp へ分離
- Tracy統計カウンタは command_generator_internal.h で翻訳単位間共有

ロジック変更なし（コード移動と include 整備のみ）
