+++ src/render/command_generator.cpp
@@ -1,4 +1,5 @@
 #include "command_generator.h"
+#include "command_generator_internal.h"
 #include "i18n.h"
 #include "layout.h"
 #include "layout_computer.h"
@@ -10,18 +11,10 @@
 #include <utility>
 
 #ifdef MENDO_USE_TRACY
-namespace {
-
-// 累積カウンタ + 直近フレーム値（UI スレッド単一前提のため非アトミック）。
-struct CmdGenStats {
-    int64_t hittest_range = 0;           // FetchHitTestMetrics (HitTestTextRange) 呼び出し
-    int64_t sel_hl_cache_hit = 0;        // SelectionHlCache ヒット
-    int64_t sel_hl_cache_miss = 0;       // SelectionHlCache ミス (HitTestTextRange を回す)
-    int64_t search_hl_rebuild = 0;       // SearchHlCache 再構築
-    int64_t search_hl_provisional = 0;   // 表セルレイアウト失効中の暫定空 rect プッシュ (次フレーム再構築待ち)
-    int64_t last_visible_node_count = 0; // 累積ではなく直近フレームのスナップショット
-};
+// 分割ファイル間で共有する統計カウンタの実体定義。
 CmdGenStats g_cmd_gen_stats;
+
+namespace {
 
 void PublishCmdGenStats() noexcept
 {
@@ -80,16 +73,6 @@ private:
 };
 
 } // namespace
-
-// DWRITE_HIT_TEST_METRICS を origin 加算付きの D2D1_RECT_F に変換する。
-static inline D2D1_RECT_F RectFromHitTest(const DWRITE_HIT_TEST_METRICS& m, float origin_x = 0.0f, float origin_y = 0.0f) noexcept
-{
-    return D2D1::RectF(
-        origin_x + m.left,
-        origin_y + m.top,
-        origin_x + m.left + m.width,
-        origin_y + m.top + m.height);
-}
 
 // MeasureNode 時に確定済みの first_line_height を優先利用する。
 // 0（未確定）なら DirectWrite に問い合わせず font_size ベースのフォールバックで済ませ、
@@ -652,316 +635,3 @@ void CommandGenerator::GenDiagramPlaceholder(DrawCommandList& cmds, float x, flo
     }
 }
 
-void CommandGenerator::EmitHighlightRects(
-    DrawCommandList& cmds,
-    IDWriteTextLayout* layout,
-    uint32_t start,
-    uint32_t length,
-    float origin_x,
-    float origin_y,
-    D2D1_COLOR_F color,
-    BrushId brush_id)
-{
-    if (!layout || length == 0) {
-        return;
-    }
-    assert(hit_test_buffer_ && "SetHitTestBuffer must be called before GenerateMdPane");
-    auto& buf = *hit_test_buffer_;
-    MENDO_COUNT_INC(g_cmd_gen_stats.hittest_range);
-    const UINT32 count = FetchHitTestMetrics(layout, start, length, buf);
-    for (UINT32 i = 0; i < count; i++) {
-        cmds.emplace_back(FillRectCmd{ RectFromHitTest(buf[i], origin_x, origin_y), color, brush_id });
-    }
-}
-
-void CommandGenerator::GenSelectionHighlight(DrawCommandList& cmds, IDWriteTextLayout* layout, uint32_t start, uint32_t length, float origin_x, float origin_y)
-{
-    EmitHighlightRects(cmds, layout, start, length, origin_x, origin_y, SELECTION_COLOR, BrushId::Selection);
-}
-
-void CommandGenerator::CollectHitTestRects(IDWriteTextLayout* layout, uint32_t start, uint32_t length, std::pmr::vector<D2D1_RECT_F>& out)
-{
-    assert(hit_test_buffer_ && "SetHitTestBuffer must be called before GenerateMdPane");
-    auto& buf = *hit_test_buffer_;
-    MENDO_COUNT_INC(g_cmd_gen_stats.hittest_range);
-    const UINT32 count = FetchHitTestMetrics(layout, start, length, buf);
-    out.reserve(out.size() + count);
-    for (UINT32 i = 0; i < count; i++) {
-        out.emplace_back(RectFromHitTest(buf[i]));
-    }
-}
-
-void CommandGenerator::GenSelectionHighlightCached(DrawCommandList& cmds, const Node& node, const NodeLayoutEntry& entry, uint32_t doc_start, uint32_t doc_length, float origin_x, float origin_y)
-{
-    auto* layout = entry.text_layout.Get();
-    if (!layout || doc_length == 0) {
-        return;
-    }
-    auto& cache = entry.ensure_selection_hl_cache();
-    // キャッシュキーは UTF-8 byte 単位 (selection 状態と直接対応)。
-    // miss 時は WideViewCache 経由で同一ノードの連続 miss でも decode を 1 回に抑える。
-    if (cache.layout_ptr != layout || cache.start != doc_start || cache.length != doc_length) {
-        MENDO_COUNT_INC(g_cmd_gen_stats.sel_hl_cache_miss);
-        cache.rects.clear();
-        const auto wr = node_wv_.WideRange(node.GetText(), doc_start, doc_length);
-        if (wr.length > 0) {
-            CollectHitTestRects(layout, wr.startPosition, wr.length, cache.rects);
-        }
-        cache.layout_ptr = layout;
-        cache.start = doc_start;
-        cache.length = doc_length;
-    }
-    else {
-        MENDO_COUNT_INC(g_cmd_gen_stats.sel_hl_cache_hit);
-    }
-    for (const auto& r : cache.rects) {
-        cmds.emplace_back(FillRectCmd{ OffsetRectF(r, origin_x, origin_y), SELECTION_COLOR, BrushId::Selection });
-    }
-}
-
-void CommandGenerator::GenSearchHighlights(DrawCommandList& cmds, const NodeLayoutEntry& entry, int node_index, float origin_x, float origin_y, int table_row, int table_col)
-{
-    if (!search_matches_ || search_matches_->empty()) {
-        return;
-    }
-
-    const std::span<const SearchMatch> matches = *search_matches_;
-    const auto range = std::ranges::equal_range(matches, node_index, {}, &SearchMatch::node_index);
-    if (range.empty()) {
-        return;
-    }
-    const size_t first_global = static_cast<size_t>(range.begin() - matches.begin());
-    const size_t node_match_count = static_cast<size_t>(range.size());
-
-    auto& cache = entry.ensure_search_hl_cache();
-    RebuildSearchHlCache(cache, entry, matches, first_global, node_match_count);
-    EmitSearchHlCommands(cmds, cache, matches, first_global, origin_x, origin_y, table_row, table_col);
-}
-
-void CommandGenerator::RebuildSearchHlCache(
-    SearchHlCache& cache, const NodeLayoutEntry& entry,
-    std::span<const SearchMatch> matches, size_t first_global, size_t node_match_count)
-{
-    // キャッシュミス時のみ HitTestTextRange を一括発行。layout 変更時は
-    // invalidate_search_hl_cache() でキャッシュ自体が破棄されており、SearchState の
-    // generation は 1 から始まるため、cache.gen == search_generation_ のみで
-    // キャッシュ有効性を完全判定できる。
-    if (cache.gen == search_generation_) {
-        return;
-    }
-
-    MENDO_COUNT_INC(g_cmd_gen_stats.search_hl_rebuild);
-    cache.rects.clear();
-    cache.rect_ends.clear();
-    cache.rect_ends.reserve(node_match_count);
-
-    for (size_t mi = first_global; mi < first_global + node_match_count; ++mi) {
-        const auto& m = matches[mi];
-        IDWriteTextLayout* l = nullptr;
-        if (m.table_row >= 0 && entry.has_table_layout()) {
-            l = entry.table_layout->GetCellLayout(static_cast<size_t>(m.table_row), static_cast<size_t>(m.table_col));
-        }
-        else if (m.table_row < 0) {
-            l = entry.text_layout.Get();
-        }
-        // m.table_row >= 0 && !has_table_layout() のケースは layout 失効中の
-        // 暫定状態で、l は nullptr のまま空 rect として記録する（次回再構築時に修復）。
-        if (l && m.length_w > 0) {
-            CollectHitTestRects(l, m.start_w, m.length_w, cache.rects);
-        }
-        else if (!l && m.table_row >= 0) {
-            MENDO_COUNT_INC(g_cmd_gen_stats.search_hl_provisional);
-        }
-        cache.rect_ends.push_back(static_cast<uint32_t>(cache.rects.size()));
-    }
-    cache.gen = search_generation_;
-}
-
-void CommandGenerator::EmitSearchHlCommands(
-    DrawCommandList& cmds,
-    const SearchHlCache& cache,
-    std::span<const SearchMatch> matches,
-    size_t first_global,
-    float origin_x,
-    float origin_y,
-    int table_row,
-    int table_col)
-{
-    // 呼び出し側（セル/ノード本体）に属する match のみ描画する。
-    const size_t node_match_count = cache.rect_ends.size();
-
-    for (size_t node_mi = 0; node_mi < node_match_count; ++node_mi) {
-        const size_t mi = first_global + node_mi;
-        const auto& m = matches[mi];
-        const bool is_here = (table_row >= 0) ? (m.table_row == table_row && m.table_col == table_col) : (m.table_row < 0);
-        if (!is_here) {
-            continue;
-        }
-
-        const uint32_t rb = (node_mi == 0) ? 0 : cache.rect_ends[node_mi - 1];
-        const uint32_t re = cache.rect_ends[node_mi];
-
-        const bool is_current = (static_cast<int>(mi) == current_match_index_);
-        const D2D1_COLOR_F color = is_current ? theme_->search_highlight_current_color : theme_->search_highlight_color;
-        const BrushId hl_brush = is_current ? BrushId::SearchHighlightCurrent : BrushId::SearchHighlight;
-        for (uint32_t k = rb; k < re; ++k) {
-            cmds.emplace_back(FillRectCmd{ OffsetRectF(cache.rects[k], origin_x, origin_y), color, hl_brush });
-        }
-    }
-}
-
-void CommandGenerator::GenTableRowBg(DrawCommandList& cmds, const TableRowGeom& g, bool is_header, bool is_even_row)
-{
-    if (is_header) {
-        cmds.emplace_back(FillRectCmd{ D2D1::RectF(g.x, g.y, g.x + g.table_width, g.y + g.row_h + g.border), theme_->code_bg_color, BrushId::CodeBg });
-    }
-    else if (is_even_row) {
-        // cached_stripe_color_ と Renderer の brushes_[TableStripe] は同じ式 (renderer_resources.cpp) で算出する。
-        cmds.emplace_back(FillRectCmd{ D2D1::RectF(g.x, g.y, g.x + g.table_width, g.y + g.row_h + g.border), cached_stripe_color_, BrushId::TableStripe });
-    }
-}
-
-void CommandGenerator::GenTableCellContent(DrawCommandList& cmds, std::string_view cell_text, const CellDrawContext& ctx)
-{
-    if (ctx.has_selection && ctx.layout) {
-        const uint32_t cell_len = static_cast<uint32_t>(cell_text.size());
-        const uint32_t ov_start = std::max(ctx.sel_start, ctx.flat_offset);
-        const uint32_t ov_end = std::min(ctx.sel_end, ctx.flat_offset + cell_len);
-        if (ov_end > ov_start) {
-            // sel_start/sel_end は UTF-8 byte offset。HitTestTextRange 用に UTF-16 へ変換する。
-            const auto wr = cell_wv_.WideRange(cell_text, ov_start - ctx.flat_offset, ov_end - ov_start);
-            if (wr.length > 0) {
-                GenSelectionHighlight(cmds, ctx.layout, wr.startPosition, wr.length, ctx.text_x, ctx.text_y);
-            }
-        }
-    }
-    if (ctx.layout) {
-        const D2D1_COLOR_F cell_color = ctx.is_header ? theme_->heading_color : theme_->text_color;
-        const BrushId cell_brush = ctx.is_header ? BrushId::Heading : BrushId::Text;
-        cmds.emplace_back(DrawTextLayoutCmd{ D2D1::Point2F(ctx.text_x, ctx.text_y), ctx.layout, cell_color, cell_brush });
-    }
-}
-
-void CommandGenerator::GenTable(
-    DrawCommandList& cmds,
-    const FrameContext& fc,
-    const Node& node, const NodeLayoutEntry& entry,
-    int node_index, float offset_x, float entry_text_top, float h_scroll_x)
-{
-    const auto* tbl = node.table_data();
-    if (!tbl || tbl->row_count == 0 || !entry.has_table_layout() || entry.table_layout->col_widths.empty()) {
-        return;
-    }
-
-    const float cell_padding = TABLE_CELL_PADDING;
-    const float border = TABLE_BORDER_WIDTH;
-    const auto& tl = *entry.table_layout;
-    const auto& selection = fc.selection;
-    const float viewport_top = fc.viewport_top;
-    const float viewport_bottom = fc.viewport_bottom;
-    const auto row_count = tbl->row_count;
-    const auto col_count = static_cast<size_t>(tbl->col_count);
-
-    const float table_width = tl.cached_table_width;
-
-    bool has_selection = selection.active && (node_index >= selection.start_node) && (node_index <= selection.end_node);
-    uint32_t sel_start = 0, sel_end = 0;
-    if (has_selection) {
-        const auto range = selection.ClampedRange(node_index, tbl->concat_text.size());
-        sel_start = range.start;
-        sel_end = range.end;
-        if (sel_end <= sel_start) {
-            has_selection = false;
-        }
-    }
-
-    float y = entry_text_top;
-    size_t bg_cursor = 0;
-
-    // 可視行帯を row_cum_y の二分探索で求め、その範囲だけループする
-    // (ApplyTableEffects / FindTableRow と同じパターン)。巨大テーブルで
-    // 毎フレーム row_count 回の continue ループが走るのを防ぐ。
-    size_t r_begin = 0;
-    size_t r_end = row_count;
-    const bool has_row_geometry = tl.row_cum_y.size() == row_count + 1;
-    if (has_row_geometry) {
-        const auto [rb, re] = tl.VisibleRowRange(viewport_top - entry_text_top, viewport_bottom - entry_text_top);
-        r_begin = rb;
-        r_end = re;
-        y = entry_text_top + tl.row_cum_y[r_begin];
-        // bg リストは追記順が乱れうるため二分探索せず、既存セマンティクス
-        // (前進スキップ) で r_begin 直前まで進める。サイズは bg 持ちセル数のみ。
-        const uint32_t first_cell = static_cast<uint32_t>(r_begin * tl.col_count);
-        while (bg_cursor < tl.cell_inline_code_bgs.size() && tl.cell_inline_code_bgs[bg_cursor].cell_index < first_cell) {
-            ++bg_cursor;
-        }
-    }
-
-    for (size_t r = r_begin; r < r_end; r++) {
-        const float row_h = (r < tl.row_heights.size()) ? tl.row_heights[r] : (theme_->font_size_body * TABLE_ROW_HEIGHT_FACTOR);
-
-        const float row_bottom = y + row_h + border;
-        if (row_bottom < viewport_top || y > viewport_bottom) {
-            y = row_bottom;
-            if (!tl.cell_inline_code_bgs.empty() && bg_cursor < tl.cell_inline_code_bgs.size()) {
-                const uint32_t next_cell = static_cast<uint32_t>((r + 1) * tl.col_count);
-                while (bg_cursor < tl.cell_inline_code_bgs.size() && tl.cell_inline_code_bgs[bg_cursor].cell_index < next_cell) {
-                    ++bg_cursor;
-                }
-            }
-            continue;
-        }
-
-        const bool is_header_row = tbl->IsHeaderRow(r);
-        GenTableRowBg(cmds, TableRowGeom{ offset_x, y, table_width, row_h, border }, is_header_row, r % 2 == 0);
-
-        // 行上部の水平線
-        cmds.emplace_back(DrawLineCmd{ D2D1::Point2F(offset_x, y), D2D1::Point2F(offset_x + table_width, y), theme_->hr_color, border, BrushId::Hr });
-
-        // 可視列のみコマンド生成、画面外は cell_text_starts で flat_offset を直接取得。
-        // 横スクロール時は画面に映る範囲が h_scroll_x だけ右にずれる。
-        const float cull_left = offset_x + fc.viewport_left + h_scroll_x;
-        const float cull_right = offset_x + fc.viewport_right + h_scroll_x;
-        float cx = offset_x + border;
-        const size_t drawn_cols = std::min(col_count, tl.col_widths.size());
-        for (size_t c = 0; c < drawn_cols; c++) {
-            const float cw = tl.col_widths[c];
-            const float col_right = cx + cw + cell_padding * 2.0f;
-            const bool col_visible = (col_right >= cull_left) && (cx - border <= cull_right);
-
-            if (col_visible) {
-                cmds.emplace_back(DrawLineCmd{ D2D1::Point2F(cx - border, y), D2D1::Point2F(cx - border, y + row_h + border), theme_->hr_color, border, BrushId::Hr });
-
-                const float text_x = cx + cell_padding;
-                const float text_y = y + cell_padding;
-                IDWriteTextLayout* cell_layout = tl.GetCellLayout(r, c);
-
-                // bg_cursor は cell_index 昇順で進む。
-                bg_cursor = GenCellInlineCodeBgs(cmds, tl.cell_inline_code_bgs, bg_cursor, static_cast<uint32_t>(tl.CellIndex(r, c)), text_x, text_y, theme_->code_bg_color);
-
-                GenSearchHighlights(cmds, entry, node_index, text_x, text_y, static_cast<int>(r), static_cast<int>(c));
-
-                const auto cell_text = tbl->GetCellText(r, c);
-                const uint32_t cell_flat = tbl->CellTextStart(r, c);
-                GenTableCellContent(cmds, cell_text, CellDrawContext{
-                    .layout = cell_layout,
-                    .text_x = text_x,
-                    .text_y = text_y,
-                    .flat_offset = cell_flat,
-                    .sel_start = sel_start,
-                    .sel_end = sel_end,
-                    .has_selection = has_selection,
-                    .is_header = is_header_row,
-                });
-            }
-
-            cx += cw + cell_padding * 2.0f + border;
-        }
-        cmds.emplace_back(DrawLineCmd{ D2D1::Point2F(offset_x + table_width, y), D2D1::Point2F(offset_x + table_width, y + row_h + border), theme_->hr_color, border, BrushId::Hr });
-        y += row_h + border;
-    }
-
-    // r_end で打ち切った場合も下端線はテーブル全体の底に置く (可視外ならクリップされる)
-    const float table_bottom = has_row_geometry ? entry_text_top + tl.row_cum_y[row_count] : y;
-    cmds.emplace_back(DrawLineCmd{ D2D1::Point2F(offset_x, table_bottom), D2D1::Point2F(offset_x + table_width, table_bottom), theme_->hr_color, border, BrushId::Hr });
-}
